requirements.txt
python-telegram-bot==21.4
requests
python-dotenv
#streamlit
asyncio
nest_asyncio
